edge-backend/app/sql_app/db_manager.py
import os

from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.orm import Session, declarative_base, sessionmaker

"""
This module configures the database connection using SQLAlchemy. It establishes settings for the database engine, session, and base model for ORM mappings.
"""

env = os.getenv("ENV")
load_dotenv()
DATABASE_URL = os.getenv("DATABASE_URL")

SQLALCHEMY_DATABASE_URL = DATABASE_URL
engine = create_engine(SQLALCHEMY_DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
Base = declarative_base()
